[Bugfix] Fix load config when using bools (#9533)
