+++ tests/data/test_config.yaml
@@ -1,3 +1,5 @@
 port: 12312
 served_model_name: mymodel
 tensor_parallel_size: 2
+trust_remote_code: true
+multi_step_stream_outputs: false
